requirements/prod.txt
django==1.7.6
django-haystack==2.4.0
celery==3.1.17
Jinja2==2.7.2
MarkupSafe==0.18
PyYAML>=3.10
gunicorn==19.1.1
requests==2.5.0
python-memcached==1.53
redis==2.10.3
django-suit==0.2.12
enum34==0.9.23
django-datetime-widget==0.9.3
django-location-field==1.5.2
django-localflavor
elasticsearch==1.6.0
MySQL-python==1.2.5
